Initial introduction of a opam-release tool
This is a tool to generate a branch (and a pull-request manually) from the opam meta data local to the repository.
Most of the arguments are inferred automatically if one:
- has a unique file named <your-project>.opam
- is using a github repository
- has their github username configured in their global .gitconfig
- has performed git remote update
- the dev-repo is set in the opam file
- opam is installed

Arguments can be supplied to fix any of the default behavior (or their absence thereof).
One can witness the inferred arguments by providing option `-s`