[Nix Action] Deploy to Cachix even for PR from forks.
Motivation

The goal is to avoid ever rebuilding the same thing twice in CI. When
a PR is submitted from a fork, if it has to build a dependency, we
don't want to waste this precious build time by discarding the build
output. And if the PR is merged as is, it shouldn't be needed to
rebuild anything once merged in the base branch (if it hasn't
changed either).

Solution

Use the `pull_request_target` event instead of the `pull_request`
event to have access to secret environment variables in all
cases. This is safe (no risk of linking the secret variables) even
with untrusted code because `nix-build` builds the code in an isolated
environment. The `pull_request_target` event ensures that the
submitter of the pull request cannot change the GitHub workflow.

Note about modified workflows

When a GitHub workflow is modified, in order to test it, we do run it
with the pull_request event. In this case, it is recommended submitted
from a local branch (not from a fork) or to have an appropriate
CACHIX_AUTH_TOKEN secret variable set in the fork.

This commit follows the equivalent coq-community/coq-nix-toolbox#55
for the standalone CI generation procedure of the Coq Nix Toolbox.